extracting currencies fragment
